# docs/api/api/openapi.yaml
openapi: 3.1.0
info:
  title: Courier Delivery Management System
  description: "\n        ## Courier Delivery Management System API\n\n        A comprehensive\
    \ API for managing courier delivery operations with real-time tracking and updates.\n\
    \n        ### Features\n        - **User Management**: Authentication and role-based\
    \ access control\n        - **Package Management**: Create, track, and manage\
    \ delivery packages\n        - **Real-time Updates**: WebSocket connections for\
    \ live notifications\n        - **Role-based Access**: Admin, Courier, Sender,\
    \ and Recipient roles\n        - **Status Tracking**: Complete package lifecycle\
    \ management\n\n        ### Authentication\n        This API uses JWT (JSON Web\
    \ Tokens) for authentication. Include the token in the Authorization header:\n\
    \        ```\n        Authorization: Bearer <your-jwt-token>\n        ```\n\n\
    \        ### User Roles\n        - **Admin**: Full system access and management\n\
    \        - **Courier**: Manage assigned packages and update delivery status\n\
    \        - **Sender**: Create packages and track deliveries\n        - **Recipient**:\
    \ View packages addressed to them\n\n        ### Getting Started\n        1. Register\
    \ a new user account\n        2. Login to receive an access token\n        3.\
    \ Use the token to access protected endpoints\n        4. Connect to WebSocket\
...
    \        "
  contact:
...
  license:
...
servers:
...
paths:
...
  /api/users/:
    get:
      tags:
      - users
      summary: List Users
      description: Get list of users (Admin only)
      operationId: list_users_api_users__get
      security:
      - HTTPBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of users to skip
          default: 0
          title: Skip
        description: Number of users to skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Maximum number of users to return
          default: 10
          title: Limit
        description: Maximum number of users to return
      - name: role
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/UserRole'
          - type: 'null'
          description: Filter by user role
          title: Role
        description: Filter by user role
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search by name or email
          title: Search
        description: Search by name or email
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
                title: Response List Users Api Users  Get
              example:
              - id: 1
                email: admin@example.com
                name: Admin User
                role: admin
                phone: '+1234567890'
                created_at: '2024-01-01T10:00:00Z'
                updated_at: '2024-01-01T10:00:00Z'
        '401':
          description: Authentication required
          content:
            application/json:
              example:
                detail: Not authenticated
        '403':
          description: Insufficient permissions
          content:
            application/json:
              example:
                detail: 'Access denied. Required role: admin'
        '404':
          description: Resource not found
          content:
            application/json:
              example:
                detail: Resource not found
        '422':
          description: Validation error
          content:
            application/json:
              example:
                detail:
                - loc:
                  - body
                  - email
                  msg: field required
                  type: value_error.missing
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                detail: Internal server error
  /api/users/me:
    get:
      tags:
      - users
      summary: Get Current User Profile
      description: Get current user's profile information
      operationId: get_my_profile_api_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - HTTPBearer: []
    put:
      tags:
      - users
      summary: Update Current User Profile
      description: Update current user's profile information
      operationId: update_my_profile_api_users_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/users/{user_id}:
    get:
      tags:
      - users
      summary: Get User by ID
      description: Get user information by ID (Admin or own profile)
      operationId: get_user_api_users__user_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - users
      summary: Update User
      description: Update user information (Admin only)
      operationId: update_user_api_users__user_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - users
      summary: Delete User
      description: Delete user account (Admin only)
      operationId: delete_user_api_users__user_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/{user_id}/role:
    put:
      tags:
      - users
      summary: Update User Role
      description: Update user role (Admin only)
      operationId: update_user_role_api_users__user_id__role_put
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/role/{role}:
    get:
      tags:
      - users
      summary: Get Users by Role
      description: Get users filtered by role
      operationId: get_users_by_role_api_users_role__role__get
      security:
      - HTTPBearer: []
      parameters:
      - name: role
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UserRole'
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
                title: Response Get Users By Role Api Users Role  Role  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/users/stats/summary:
    get:
      tags:
      - users
      summary: Get User Statistics
      description: Get user statistics summary (Admin only)
      operationId: get_user_stats_api_users_stats_summary_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserResponse:
      properties:
        email:
          type: string
          format: email
          title: Email
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        role:
          $ref: '#/components/schemas/UserRole'
        phone:
          anyOf:
          - type: string
            maxLength: 20
          - type: 'null'
          title: Phone
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - email
      - name
      - role
      - id
      - created_at
      - updated_at
      title: UserResponse
    UserRole:
      type: string
      enum:
      - admin
      - courier
      - sender
      - recipient
      title: UserRole
    UserRoleUpdate:
      properties:
        role:
          allOf:
          - $ref: '#/components/schemas/UserRole'
          description: New role for the user
      type: object
      required:
      - role
      title: UserRoleUpdate
    UserUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Name
        phone:
          anyOf:
          - type: string
            maxLength: 20
          - type: 'null'
          title: Phone
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login endpoint
security:
- BearerAuth: []
...